news-consumer/requirements.txt
openai==0.28
pymongo
confluent-kafka
python-dateutil
fastapi
uvicorn